snmp-ups: fix regression on Eaton ePDU
Fix a regression that caused a mis-determination of the SNMP base
OID index (0 or 1, should be 1). This in turn caused a mis-iteration
over the outlets, from 0 to N-1 instead of 1 to N, which resulted in
the missing Nth outlet (last outlet of the PDU). This also caused some
data refresh issues

Signed-off-by: Arnaud Quette <ArnaudQuette@eaton.com>